Move analytics tests to integration tests and add new test case
